services/extractor/requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
pandas==2.2.2
pyarrow==17.0.0
numpy==2.1.2
statsmodels==0.14.3
gunicorn==21.2.0
db-dtypes==1.3.0
